Split CFG creation into multiple files
