+++ src/test/scala/io/shiftleft/fuzzyc2cpg/passes/CfgCreationPassTests.scala
@@ -8,7 +8,7 @@ import io.shiftleft.semanticcpg.language._
 
 import scala.jdk.CollectionConverters._
 import io.shiftleft.codepropertygraph.generated.nodes
-import io.shiftleft.fuzzyc2cpg.passes.CfgCreatorForMethod.{AlwaysEdge, CaseEdge, CfgEdgeType, FalseEdge, TrueEdge}
+import io.shiftleft.fuzzyc2cpg.passes.cfgcreation.Cfg.{AlwaysEdge, CaseEdge, CfgEdgeType, FalseEdge, TrueEdge}
 
 class CfgCreationPassTests extends WordSpec with Matchers {
